Rename Schema -> Allocator in file names
