+++ Code/Framework/AzCore/AzCore/Memory/HphaAllocator.h
