+++ Code/Framework/AzCore/Tests/Memory/AllocatorBenchmarks.cpp
@@ -11,9 +11,9 @@
 #include <AzCore/PlatformIncl.h>
 #include <AzCore/IO/SystemFile.h>
 #include <AzCore/RTTI/TypeInfo.h>
-#include <AzCore/Memory/HphaSchema.h>
+#include <AzCore/Memory/HphaAllocator.h>
 #include <AzCore/Memory/OSAllocator.h>
-#include <AzCore/Memory/PoolSchema.h>
+#include <AzCore/Memory/PoolAllocator.h>
 #include <AzCore/Memory/SystemAllocator.h>
 #include <AzCore/std/containers/array.h>
 #include <AzCore/std/containers/vector.h>
